f16 block scales end to end (~10% fewer weight bytes per token)
Keep the GGUF's native f16 scales instead of widening to f32: packed two per
u32, decoded with unpack2x16float in every matvec variant (legacy, coop,
batched). Requantized tensors round their computed scales to f16 first
(matching the GGUF spec, like llama.cpp). CPU embed-row dequant and
self-test quantizers updated. Suite 11/11 green; streamed loader still
bit-identical; GB10 27B decode 8.6 -> 9.0 tok/s.

+++ engine/test_stream_deno.js
@@ -19,7 +19,7 @@ for (const [name, info] of Object.entries(G.tensors)) {
   const ref = (info.ggmlType === GGML_Q4_0 ? q4Repack : q8Repack)(info, await slice(info));
   const e = await streamEntryToGPU(device, info, openRange, { staging: 1 << 20 });
   const qs = await readBack(e.gpu.qs, ref.qs.byteLength);
-  const sc = new Float32Array((await readBack(e.gpu.sc, ref.scales.byteLength)).buffer);
+  const sc = new Uint32Array((await readBack(e.gpu.sc, ref.scales.byteLength)).buffer);
   let bad = 0; for (let i = 0; i < ref.qs.length; i++) if (qs[i] !== ref.qs[i]) { bad++; break; }
   for (let i = 0; i < ref.scales.length; i++) if (sc[i] !== ref.scales[i]) { bad++; break; }
   if (bad) { console.log("MISMATCH", name); Deno.exit(1); }
